+++ R/zzz.R
@@ -31,12 +31,12 @@ NULL
 
 .onAttach <- function(libname, pkgname) {
   vers <-  "1.0.0"
-  packageStartupMessage("|=======================================================|")
-  packageStartupMessage("| Tools for Plant Image Analysis (pliman ", vers,  ")         |")
-  packageStartupMessage("| Author: Tiago Olivoto                                 |")
-  packageStartupMessage("| Type 'vignette('pliman_start')' for a short tutorial  |")
-  packageStartupMessage("| Visit 'https://bit.ly/pliman' for a complete tutorial |")
-  packageStartupMessage("|=======================================================|")
+  packageStartupMessage("|==========================================================|")
+  packageStartupMessage("| Tools for Plant Image Analysis (pliman ", vers,  ")            |")
+  packageStartupMessage("| Author: Tiago Olivoto                                    |")
+  packageStartupMessage("| Type 'vignette('pliman_start')' for a short tutorial     |")
+  packageStartupMessage("| Visit 'http://bit.ly/pkg_pliman' for a complete tutorial |")
+  packageStartupMessage("|==========================================================|")
   check_ebi()
 }